Remove command from main application menu
Since most packages add their commands to the command palette, I find
that I always look for this command first in the command palette, only
to then realize that it's not there; it's in the application menu
instead. Well, let's change that.

This commit removes the command from the application menu.

36f7748 adds this command to the command pallete.